+++ apps/server/app.js
@@ -2,10 +2,11 @@ var express = require("express");
 var path = require("path");
 var cookieParser = require("cookie-parser");
 var logger = require("morgan");
+const cors = require("cors");
 
-var indexRouter = require("./routes/index");
+const { router: uploadRouter } = require("./upload");
 
-var app = express();
+const app = express();
 
 app.use(logger("dev"));
 app.use(express.json());
@@ -13,6 +14,11 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
-app.use("/", indexRouter);
+app.use(
+  cors({
+    origin: /localhost/,
+  })
+);
+app.use("/", uploadRouter);
 
 module.exports = app;