+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- server: /upload endpoint to upload single files to server disk storage
+
 ## [0.1.0] - 2022-05-11
 
 ### Added